# backend/src/astro_canvas/server/openapi/openapi.json
{
  "components": {
    "schemas": {
      "BatchBinding": {
        "description": "One table column feeding one node param (``layouts.batch.columns`` entry).",
        "properties": {
          "column": {
            "title": "Column",
            "type": "string"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "param": {
            "title": "Param",
            "type": "string"
          }
        },
        "required": [
          "node",
          "param",
          "column"
        ],
        "title": "BatchBinding",
        "type": "object"
      },
      "BatchCollect": {
        "description": "One node output gathered into the results table.",
        "properties": {
          "node": {
            "title": "Node",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "string"
          },
          "prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Column prefix; defaults to none for a single collect and '<node>.' when several outputs are gathered.",
            "title": "Prefix"
          }
        },
        "required": [
          "node",
          "port"
        ],
        "title": "BatchCollect",
        "type": "object"
      },
      "BatchInfo": {
        "description": "A batch's state; ``results`` is filled in as rows finish.",
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "finished": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished"
          },
          "n_rows": {
            "title": "N Rows",
            "type": "integer"
          },
          "results": {
            "$ref": "#/components/schemas/BatchResults"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/BatchRowInfo"
            },
            "title": "Rows",
            "type": "array"
          },
          "started": {
            "title": "Started",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "workflow_id",
          "status",
          "n_rows",
          "counts",
          "started",
          "rows",
          "results"
        ],
        "title": "BatchInfo",
        "type": "object"
      },
      "BatchRequest": {
        "description": "Rows plus the spec; omitting ``spec`` uses the document's ``layouts.batch``.",
        "properties": {
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BatchSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "BatchRequest",
        "type": "object"
      },
      "BatchResults": {
        "description": "The results grid: one row per input row, JSON-safe so it can travel over REST.",
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "title": "BatchResults",
        "type": "object"
      },
      "BatchRowInfo": {
        "properties": {
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "state": {
            "enum": [
              "pending",
              "queued",
              "running",
              "done",
              "error",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "index",
          "state"
        ],
        "title": "BatchRowInfo",
        "type": "object"
      },
      "BatchSpec": {
        "description": "What to vary and what to keep from a batch run.",
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/BatchBinding"
            },
            "title": "Bindings",
            "type": "array"
          },
          "collect": {
            "items": {
              "$ref": "#/components/schemas/BatchCollect"
            },
            "title": "Collect",
            "type": "array"
          },
          "continue_on_error": {
            "default": true,
            "title": "Continue On Error",
            "type": "boolean"
          },
          "max_workers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Rows in flight; default cpu-1 with expensive nodes, else 8.",
            "title": "Max Workers"
          }
        },
        "title": "BatchSpec",
        "type": "object"
      },
...
      "BundleExportRequest": {
        "description": "``POST /api/bundles/export``.",
        "properties": {
          "dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Workspace folder; defaults to ``bundles``.",
            "title": "Dir"
          },
          "embed_inputs_max_mb": {
            "default": 200,
            "minimum": 0.0,
            "title": "Embed Inputs Max Mb",
            "type": "integer"
          },
          "include_figures": {
            "default": true,
            "title": "Include Figures",
            "type": "boolean"
          },
          "include_outputs": {
            "default": "leaves",
            "enum": [
              "leaves",
              "all",
              "none"
            ],
            "title": "Include Outputs",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id"
        ],
        "title": "BundleExportRequest",
        "type": "object"
      },
...
      "BundleLock": {
        "description": "Exactly what ran, so an import can say whether this machine can reproduce it.",
        "properties": {
          "app_version": {
            "default": "0.1.0a0",
            "title": "App Version",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "string"
          },
          "packs": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Packs",
            "type": "object"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "python": {
            "title": "Python",
            "type": "string"
          },
          "requirements": {
            "items": {
              "type": "string"
            },
            "title": "Requirements",
            "type": "array"
          }
        },
        "title": "BundleLock",
        "type": "object"
      },
      "BundleManifest": {
        "description": "What ``export_bundle`` wrote (also what ``inspect_bundle`` reads back).",
        "properties": {
          "bytes": {
            "default": 0,
            "title": "Bytes",
            "type": "integer"
          },
          "code_hashes": {
            "items": {
              "type": "string"
            },
            "title": "Code Hashes",
            "type": "array"
          },
          "figures": {
            "items": {
              "type": "string"
            },
            "title": "Figures",
            "type": "array"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputRef"
            },
            "title": "Inputs",
            "type": "array"
          },
          "lock": {
            "$ref": "#/components/schemas/BundleLock"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outputs": {
            "items": {
              "type": "string"
            },
            "title": "Outputs",
            "type": "array"
          },
          "path": {
            "default": "",
            "description": "Workspace-relative path of the ``.acw``.",
            "title": "Path",
            "type": "string"
          },
          "skipped_outputs": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Outputs",
            "type": "array"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "name"
        ],
        "title": "BundleManifest",
        "type": "object"
      },
...
      "ExportRequest": {
        "description": "Which outputs to write, and where under the workspace to put them.",
        "properties": {
          "dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Workspace-relative folder; defaults to ``exports/<workflow name>``.",
            "title": "Dir"
          },
          "overwrite": {
            "default": true,
            "title": "Overwrite",
            "type": "boolean"
          },
          "refs": {
            "description": "``'<node>.<port>'`` outputs to write.",
            "items": {
              "type": "string"
            },
            "title": "Refs",
            "type": "array"
          }
        },
        "title": "ExportRequest",
        "type": "object"
      },
      "ExportResult": {
        "properties": {
          "dir": {
            "title": "Dir",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ExportedFile"
            },
            "title": "Files",
            "type": "array"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/SkippedExport"
            },
            "title": "Skipped",
            "type": "array"
          }
        },
        "required": [
          "dir"
        ],
        "title": "ExportResult",
        "type": "object"
      },
      "ExportedFile": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "format": {
            "enum": [
              "csv",
              "npz",
              "json"
            ],
            "title": "Format",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "path",
          "bytes",
          "format"
        ],
        "title": "ExportedFile",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InputRef": {
        "description": "One file a workflow reads, as ``inputs/refs.json`` records it.",
        "properties": {
          "blake3": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blake3"
          },
          "bytes": {
            "default": 0,
            "title": "Bytes",
            "type": "integer"
          },
          "embedded": {
            "default": false,
            "title": "Embedded",
            "type": "boolean"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          },
          "missing": {
            "default": false,
            "description": "Not present in the exporting workspace.",
            "title": "Missing",
            "type": "boolean"
          },
          "param_ref": {
            "description": "``'<node>.<param>'`` that names the file.",
            "title": "Param Ref",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "param_ref",
          "path"
        ],
        "title": "InputRef",
        "type": "object"
      },
...
      "SkippedExport": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "enum": [
              "no_output",
              "bad_ref",
              "not_written"
            ],
            "title": "Reason",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "reason",
          "message"
        ],
        "title": "SkippedExport",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Astro Canvas",
    "version": "0.1.0a0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bundles/export": {
      "post": {
        "description": "Pack a workflow, its inputs, its finished outputs and its provenance into a ``.acw``.",
        "operationId": "create_bundle_api_bundles_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Bundle",
        "tags": [
          "bundles"
        ]
      }
    },
...
    "/api/workflows/{workflow_id}/batch": {
      "post": {
        "description": "Run the stored document once per row; returns immediately with the batch record.",
        "operationId": "start_batch_api_workflows__workflow_id__batch_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Batch",
        "tags": [
          "batch"
        ]
      }
    },
    "/api/workflows/{workflow_id}/batch/{batch_id}": {
      "get": {
        "description": "The batch's per-row states and the results assembled so far.",
        "operationId": "get_batch_api_workflows__workflow_id__batch__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch",
        "tags": [
          "batch"
        ]
      }
    },
    "/api/workflows/{workflow_id}/batch/{batch_id}/cancel": {
      "post": {
        "description": "Stop queued rows and cancel the ones in flight (expensive nodes have their worker killed).",
        "operationId": "cancel_batch_api_workflows__workflow_id__batch__batch_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Batch",
        "tags": [
          "batch"
        ]
      }
    },
    "/api/workflows/{workflow_id}/exports": {
      "post": {
        "description": "Write the named outputs into the workspace and return their paths.",
        "operationId": "create_export_api_workflows__workflow_id__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Export",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}